requirements.txt
requests
ShopifyAPI
